cherry-picked FOGL-5001 to 1.9.0RC

+++ C/plugins/storage/sqlite/common/include/readings_catalogue.h
@@ -73,7 +73,7 @@ public:
 	tyReadingReference getReadingReference(Connection *connection, const char *asset_code);
 	bool          attachDbsToAllConnections();
 	std::string   sqlConstructMultiDb(std::string &sqlCmdBase, std::vector<std::string>  &assetCodes);
-	int           purgeAllReadings(sqlite3 *dbHandle, const char *sqlCmdBase, char **errMsg = NULL, unsigned int *rowsAffected = NULL);
+	int           purgeAllReadings(sqlite3 *dbHandle, const char *sqlCmdBase, char **errMsg = NULL, unsigned long *rowsAffected = NULL);
 
 	bool          connectionAttachAllDbs(sqlite3 *dbHandle);
 	bool          connectionAttachDbList(sqlite3 *dbHandle, std::vector<int> &dbIdList);
